+++ src/XRP/wallet.ts
@@ -115,6 +115,8 @@ class Wallet {
   /**
    * Generate a new hierarchical deterministic wallet from a seed and derivation path.
    *
+   * @deprecated Please use methods on `WalletFactory` to generate wallets instead.
+   *
    * @param seed - The given seed for the wallet.
    * @param derivationPath - The given derivation path to use. If undefined, the default path is used.
    * @param test - Whether the address is for use on a test network, defaults to `false`.
@@ -131,17 +133,17 @@ class Wallet {
       return undefined
     }
 
-    const publicKey = Wallet.hexFromBuffer(node.publicKey)
-    const privateKey = Wallet.hexFromBuffer(node.privateKey)
+    const publicKey = Utils.hexFromBuffer(node.publicKey)
+    const privateKey = Utils.hexFromBuffer(node.privateKey)
     return new Wallet(publicKey, `00${privateKey}`, test)
   }
 
   /**
    * Generate a new wallet from the given seed.
    *
-   * @deprecated Please use `WalletFactory` instead.
+   * @deprecated Please use methods on `WalletFactory` to generate wallets instead.
    *
-   * @param seed - The given seed for the wallet.
+   * @param seed - A hex encoded seed string.
    * @param test - Whether the address is for use on a test network, defaults to `false`.
    * @returns A new wallet from the given seed, or undefined if the seed was invalid.
    */
@@ -155,10 +157,6 @@ class Wallet {
     } catch {
       return undefined
     }
-  }
-
-  private static hexFromBuffer(buffer: Buffer): string {
-    return buffer.toString('hex').toUpperCase()
   }
 
   /**