+++ CHANGELOG.md
@@ -11,10 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - A new class `WalletFactory` encapsulates functionality for creating a `Wallet` object from a public and private key.
 - A new class `WalletFactory` encapsulates functionality for creating a `Wallet` object from a seed.
+- A new class `WalletFactory` encapsulates functionality for creating a `Wallet` object from a seed and derivation path.
 
 ### Deprecated
 
-- The `walletFromSeed` method of the `Wallet` class is deprecated. Please use the new `WalletFactory` to generate `Wallet`s from seeds.
+- The `walletFromSeed` method of the `Wallet` class is deprecated. Please use the new `WalletFactory` class to generate `Wallet`s from seeds.
+- The `generateHDWalletFromSeed` method of the `Wallet` class is deprecated. Please use the new `WalletFactory` class to generate `Wallet`s from seeds.
 
 ## [5.0.4] - 2020-05-15